+++ recipes/leftbar/general_row/general_row.stories.js
@@ -8,6 +8,7 @@ import {
   LEFTBAR_GENERAL_ROW_TYPES,
   LEFTBAR_GENERAL_ROW_CONTACT_CENTER_COLORS,
 } from '@/recipes/leftbar/general_row/general_row_constants';
+import { LEFTBAR_GENERAL_ROW_ICON_SIZES } from './general_row_constants';
 
 // Default Prop Values
 export const argsData = {
@@ -15,6 +16,7 @@ export const argsData = {
   callButtonTooltip: 'Call',
   dndTextTooltip: 'Do not Disturb',
   description: 'Description',
+  iconSize: '300',
   onClick: action('click'),
   onCall: action('call'),
 };
@@ -31,6 +33,12 @@ export const argTypesData = {
     options: Object.values(LEFTBAR_GENERAL_ROW_TYPES),
     control: {
       type: 'select',
+    },
+  },
+  iconSize: {
+    control: {
+      type: 'select',
+      options: Object.values(LEFTBAR_GENERAL_ROW_ICON_SIZES),
     },
   },
   color: {