+++ recipes/leftbar/general_row/leftbar_general_row_icon.vue
@@ -2,7 +2,7 @@
   <dt-icon
     v-if="isIconType"
     :name="getIconName"
-    size="300"
+    :size="iconSize"
   />
   <div
     v-else-if="isContactCenterType"
@@ -40,6 +40,11 @@ export default {
     color: {
       type: String,
       default: null,
+    },
+
+    iconSize: {
+      type: String,
+      default: '300',
     },
   },