+++ recipes/list_items/contact_info/contact_info.stories.js
@@ -114,6 +114,7 @@ const VariantsTemplate = (args, { argTypes }) => createTemplateFromVueFile(
 export const Default = DefaultTemplate.bind({});
 Default.args = {
   avatarSrc: avatarImage,
+  avatarSeed: 'seed',
   avatarInitials: 'JL',
   avatarColor: 'base',
   avatarSize: 'lg',